Added student_ID to Profile
So now each user has an account (user) which automatically has a profile which has an attribute called student_ID which is a integer field.
Current bug is that the data doesn't seem to be showing up correctly in the view of the evaluations.  Will do further expierementation to isolate the issue and remedy it.
Possible that is an error in request and not data storage.  But data storage issue is a possiblity.
Also, the admin only privelages have yet to be solved.
Furthermore, the admin only page has yet to be activated.  Will work upon improving this function soon.